requirements.txt
Pillow
numpy<2.0.0
pandas==1.4.3
tensorflow==2.12.0
streamlit